WIP initial implementation of item fetch api call #1369
This is the first step in the implementation of the API call. Right
now the call returns in plain JSON format and I am assuming that the
calls will be authenticated with a user/password basic authorization.

Next steps include modifying the response format to an RDF serialization
format and explore using Token-based authentication.