Change item show view to render n3 graph format
This is the continuation of the work for presenting the items in n3
format for consumption by pushmi_pullyu.

The next step is to move the functionality of creating and exporting
a graph to a concern so we may use that code for show methods for items,
theses, and others if needed.