Rails/FilePath default changed to slashes
In the gem bump for rubocop-rails to 2.4.1 the default EnforcementStyle
changed from arguments to slashes.  That means that
`Rails.root.join('config', 'controlled_vocabularies', '*.yml')` becomes
`Rails.root.join('config/controlled_vocabularies/*.yml')`.  It was
[previously thought](https://github.com/rubocop-hq/rubocop-rails/pull/128#issuecomment-530671304)
that the first way was OS agnostic but it turns out that Rails.root is a
Pathname instance which accounts for this.  Now the slash is prefered for
fewer strings/arguments.