Refactor AIP API controllers to use a super class
Moved the logic from the items controller to the base class
AIPBaseController so we can serve the information required by PMPY
from the different objects we may serve from jupiter.

This commit has problems with obtaining the rdf annotations from files.
ActsAsRdfable.add_annotation_bindings! is not modifying the class to
include the rdf_annotations method, still need to figure out why that is
happening.