Replace hardcoded strings on aip api controllers
As part of the WIP strings were hardcoded in the AIP API controller to
define the namespaces that are included for each metadata file. These
harcoded strings have now been changed to use the strings from the
rdf-vocab gem and terms.yml file.

The harcoded strings represented namespaced that need to be repalced
after further research is done by the metadata team.