Add AIP API version 1
The AIP API will be the communication layer between Jupiter phase 2 an
the PushmiPullyu client which creates the AIP file for long term
archival. AIP was built using the information provided from fedora, as
we move away from fedora and onto PostgreSQL we need to provide an
alternative.

This API provides the information required for the AIP file in a simpler
manner avoiding fedora's overhead.