Address issues from PR #1441
Add model route constraint

A route constraint will bounce a request for an ubnknown model faster
than catchin it on the AIP API controller.

Change AIP API to be private

The AIP API will be closed off from the public and will only be used
by admins for maintanence and the PMPY client.

Rename model and asset references to entity

There was a discrepancy in the naming conventions for parameters,
methods, and variables throughout the AIP API that referenced objects
in the system.

References to these objects will be named entity in order to avoid
naming issues with rails commonly used words and OOP definitions.

Improve way helper module is loaded

Move module from concerns directory to a support directory and handle
it as a helper. Require it only when needed.

Add AIP API file path entry point

The PMPY client will not download files directly through the API,
instead it will fetch the file path and access the file directly on the
system.

This implementation provides the path for a single fileset, this can be
changed to provide the paths for all files to lower calls made to server

Add json schema validator to entity tests

Remove previous place holder test and now properly check for json
response format for Item and Thesis entities.

Add tests checking if files on entity actually exist