address CVE warnings (#1516)
While these are all security updates on the whole they don't represent vectors that would affect us.

    bump omniauth from 1.9.0 to 1.9.1 PR#1513 -- already addressed
    bump puma from 4.3.1 to 4.3.3 CVE#2020-5247 -- not used in production
    bump loofah from 2.3.0 to 2.4.0 CVE#2019-15587 -- no republished SVG elements
    bump nokogiri from 1.10.4 to 1.10.9 CVE#2019-13117
    bump brakeman from 4.7.0 to 4.8.0 CVE#2019-18409 -- local user vector

Also add `noopener noreferrer` when opening a link in a new tab to protect against a vulnerability. When opening a link in a new tab without setting rel: "noopener noreferrer", the new tab can control the parent tab's location. For example, an attacker could redirect to a phishing page.