Add safe migration for system account
As we add a system user from a migration we need to assume the user may
already exist. This approch will provide a new/existing user a password
the application is expectin. On a migration removal we don't do anything
in case the system user account was in fact there.