Add missing RDF.type predicate
The basic RDF.type predicate (a) was missing. Currently the values are
added manually based on the definitions by the metadata team.

Change n3 render method showing prefixes

Fix fixity report

The initial implementation was showing the information incorrectly by
relating all of the fixity information to the file uuid.

The resulting output now shows the relation of the file uuid to the
fixity event, and separately shows the fixity information.

Treat predicate values as independent values

Predicate values need to be added individually when building the graph
so the show up correctly and not as concatenated values. For example
predicate <http://purl.org/dc/elements/1.1/subject> is

"Pictures", "Randomness", "Unicorn"

and not concatenated values like

"Pictures, Randomness, Unicorn"

Add access rights predicate

The predicate http://purl.org/dc/terms/accessRights had not been added
before. It now shows up with the value of the visibility column.

Add predicate memberOf for entity

The predicate http://pcdm.org/models#memberOf had not been added. At
this point the value is added manually on the aip_controller listing
the valyes for member_of_paths. This should be changed to a meaningful
value showing the url for the collection.

Remove unneded rdf configuration

Community and Collection models included and old version of
configuration for rdf annotations.

Add Collections controller for AIP API

The entity graph shows the url for the collection information. The url
on its own does not have meaning and the metadata team is interested in
saving the collection and community information on the AIP.

This commit adds the most basic implementation to render the collection
information.

Add radioactive Item to seeds

In order to test the output from the AIP API we need to add entities
that have values for all available predicates and make sure the
resulting graph has these values.