add migration to fix concatenated subjects (part 1)
This will take a subject like ["Cat, Dog, Rabbit"] and make it ["Cat", "Dog", "Rabbit"] in a new deconcatenated_subject column.

We're doing this in two parts to make data cleanup a little more sane than edits in production.

This assumes that at some point in the future this new column will be reviewed and replace the Item's subject.