update UAL logo
As of October 23, we are now the University of Alberta Library.

About this Dale Askey said, "I welcome having a name that accurately reflects our identity and our mission
to serve the University of Alberta. The ability that all of you demonstrate to pull together to
instigate, organize, and develop myriad cross-organizational activities is one of our great strengths,
now reflected in our name."