WIP
Keep order for creator predicate

Values for the creator predicate need to be kept in order. In this
iplementation we are keeping the order for the Item entity by storing it
on the authorList predicate using an RDF list (which maintains order).

Add article to seeds that has published_status

Item needs to be of article type in order to show the published_status
value.

Add missing predicates and seed values

There were 2 reasons why the AIP files were missing information:

1. The seeds did not have the necessary values
2. The values were in the database but the predicates were not set up

Add predicates for existing date information

The predicates for 'date ingested' and 'record creacted in jupier' had
not been added even though the information was already on the database.

Also, add missing predicate fixtures for unit testing.