Added active storage queues to sidekiq configuration
When I was depositing items I was finding that a active_storage_analysis queue was filling up but not being processed.

https://mikerogers.io/2019/06/06/rails-6-sidekiq-queues.html describes some new queue names for ActiveStorage.

I've added the ones that make sense to me.

Note that this doesn't resolve the "This file is processing and will be available shortly" message for deposited items like I had hoped it would.