centralize abstration for thumbnail generation
Similar logic appeared in Depositable, DraftProperties, Community and the thumbnail/feature_image partials. By refactoring
the thumbnail_path into a helper and being able to call thumbnail_file on each instance we can centralize the logic and
have each class only deal with the things that concern them.

In investigating this I learned a few new things.  ImageMagick is required for variant on images, but a pdf utility like
Poppler is required to preview PDFs and ffmpeg is required to preview video. All three a requirements for our thumbnails
to work correctly.

I tried a few different things in order to test the helper.  I tried creating a dummy @logo which implemented only the
surface area that the thumbnail_path touched.  With this approach I got stuck on the rails_representation_path with the
pdf and images case.  Then I tried creating fixtures (turns out that active_storage fixtures are supported out of the box)
for these objects.  I ran into a similar road block because the fixtures wouldn't respond to variant and preview. Finally
I tried a hybrid approach using real objects for the pdf and image and fixtures for the error cases. Success!

I would have liked to also test that a useful warning was logged in the error cases but couldn't find a approach that worked
for me short of using the minitest-logger gem.  Maybe that's the best way forward.