centralize abstraction for thumbnail generation (#1681)
Similar logic appeared in Depositable, DraftProperties, Community and the thumbnail/feature_image partials. By refactoring the thumbnail_path into a helper and being able to call thumbnail_file on each instance we can centralize the logic and have each class only deal with the things that concern them.

In investigating this I learned a few new things.  ImageMagick is required for variant on images, but a pdf utility like Poppler is required to preview PDFs and ffmpeg is required to preview video. All three a requirements for our thumbnails to work correctly. https://guides.rubyonrails.org/active_storage_overview.html#previewing-files  Github Actions has ImageMagick installed by default, but not the pdf or video utilities: https://github.com/actions/virtual-environments/blob/master/images/linux/Ubuntu2004-README.md.  I allowed that test to be skipped if the environment doesn't have the dependency.

I tried a few different things in order to test the helper.  What finally worked was creating fixtures (turns out that active_storage fixtures are supported out of the box) for these objects.

I would have liked to also test that a useful warning was logged in the error cases but couldn't find a approach that worked for me short of using the minitest-logger gem.  I'll leave that for future me.