add some error handling
We want to wrap this in the ability to catch an exception if the file doesn't exist, and then not do the rest of the migration and carry on. That way we can delete the file ERA_subject_issue.txt once this is all fixed instead of having to keep it around forever to avoid breaking migrations.