add migration to fix concatenated subjects (part 1) (#1627)
This will take a subject like ["Cat, Dog, Rabbit"] and make it ["Cat", "Dog", "Rabbit"] in a new deconcatenated_subject column.

We're doing this in two parts to make data cleanup a little more sane than edits in production.

This assumes that at some point in the future this new column will be reviewed and replace the Item's subject.

We want to wrap this in the ability to catch an exception if the file doesn't exist, and then not do the rest of the migration and carry on. That way we can delete the file ERA_subject_issue.txt once this is all fixed instead of having to keep it around forever to avoid breaking migrations.