docker image can be built and deployed
This is if anyone wants to deploy this branch as a docker image and run the `rake jupiter:get_me_off_of_fedora`

A bump to the built with bundler version in Gemfile.lock required a newer version of bundler. Fixed this in the Dockerfile.deployment

Changed the predeploy script to look to this branch for getting the docker-compose with the correct docker image.

Bumped the version to 2.0.0.pre so that when/if this is deployed we'll know which version is running.