docker image can be built and deployed
A bump to the "built with bundler" version in Gemfile.lock required a newer version of bundler. Fixed this in the Dockerfile.deployment

Changed the predeploy script to look to this branch for getting the docker-compose with the correct docker image.

When assets:precompile happens Redis may not be running so ignore the Oaisys initializer that tries to create a new redis connection.

Added webpacker (to replace sprockets) so needed some tweaks to work with the new /public/packs setup.

Bumped the version to 2.0.0.pre so that when/if this is deployed we'll know which version is running.