fix serving assets
Previously we were using sprockets which precompiled the assets into a `public/assets` directory.  Moving to webpacker these assets are now precompiled into a `public/packs` directory and the nginx rule must change to `packs`.  However in order to serve the static assets from the `public` directory we need to make these available to nginx -- hence the change to the docker volume.