docker image can be built and deployed (#1761)
A bump to the "built with bundler" version in Gemfile.lock required a newer version of bundler. Fixed this in the Dockerfile.deployment

Changed the predeploy script to look to this branch for getting the docker-compose with the correct docker image.

When assets:precompile happens it loads the application including all initializers.  The Oaisys initializer would throw an error if a connection to Redis wasn't available -- exactly the situation when precompiling assets for docker.  This was fixed upstream hence the version bump here.

With the integration_postmigration branch we're moving away from Fedora which results in an orphaned container.  We use the flag to clean it up.

Previously we were using sprockets which precompiled the assets into a `public/assets` directory.  Moving to webpacker these assets are now precompiled into a `public/packs` directory and the nginx rule must change to `packs`.  However in order to serve the static assets from the `public` directory we need to make these available to nginx -- hence the change to the docker volume.