UAT VIRTUAL_HOSTS configuration on just the containers that need it
Sometimes when containers were updated by watchtower the VIRTUAL_HOSTS values in the .env_deployment files would apply to several containers.  This mean that sometimes it would try to route to the web container directly instead of nginx (and not work).

There is still a gotcha with solr.era.uat.library.ualberta.ca being redirected to https by HSTS.  To get around this try visiting in a private window or "Clear Cookies and Site Data".