rearrange the peel routes
First this makes digitization a scope rather than a namespace.  `digitization.library.ualberta.ca/book` vs `digitization.library.ualberta.ca/digitization/book`.  This changed the url helpers which were then updated to reflect the change.

Adding seeds and testing them illustrated that the order of the routes matters for how the redirects will be resolved for maps.  If the maps resource is first, everything is assumed to be `id`.  If we put the peel redirect first and add a constraint then this will filter those maps and the resource map will catch the rest.