Nokogiri::XML::ParseOptions with the NONET flag turned off
We parse schemas with Nokogiri to test valid xml in our tests (which is why the tests are failing). This appears to be the only place we do this so this security patch is low impact/low severity for us. We trust our fixture schemas and can apply the workaround described in the Github Advisory Database.

https://github.com/advisories/GHSA-vr8q-g5c7-m54m