modelling for folk fest metadata
The first set of digitization content will be the Folk Fest Programs. This set was chosen because it is small (there are 36) and well described. Before we can ingest them we need to do some modelling.

This is the first pass.  There are more details that still need to be implementated with regard to EDTF encoding.

https://docs.google.com/spreadsheets/d/1HdFwQYQV5jjDhACQgGnlBc67oOHtfAZH0iyEt7EfGnY/edit?ts=5fc95386#gid=1058768873

https://docs.google.com/spreadsheets/d/1WbW2lSdSCjKtQGkdu_mTPXxE0-o2TCmL36kupNicIng/edit?ts=5fc95397&pli=1#gid=0