migration to rename deconcatenated subjects to subjects (#1931)
## Context
In the first part of this data clean up we created a new column `deconcatenated_subject` and copied all subjects here, cleaning up the ones we've identified.

Now we want to replace the subjects column with these values.

#1449 #1627 

## What's New

We remove the subject column and rename `deconcatenated_subject` to `subject`.  Also clean up the ERA_subjects_issue.txt file and reference in the previous migration.