Don't send failures to SessionController in development environment (#2121)
Since bumping omniauth from 1.9.1 to 2.0.1 in our application all development errors have become failures. For example I purposefully introduce an error to a view. It is caught by omniauth and calls fail which means that we can no longer use better_errors to debug.

Opened a ticket with omniauth to see if they have any suggestions.