system test for facet badge categories
I had tried this as a system test.  The test had errors and I thought that Flipper wasn't working.  Turns out that invoking the system test individually was what was causing the failure.  The system test I had written wasn't correct and perhaps that contributed to my issue as well.