Labels for active facet badges (#2115)
Applied Filters could show more information. Such as the facet they belong to. Currently active facets are jumbled together.

Flipper on :facet_badge_category_name

Needed to add value and count to the RangeFacetResult so that it had the same interface in the Decorator. Also needed to add the category name to the interface so that we could use it in the Decorator.  Probably want to add a facet superclass with the common interface.