add rdau vocab to terms for rdau:extent.en
RDF::Vocab::RDAU.P60550 is "http://rdaregistry.info/Elements/u/P60550" which is purposefully opaque according to this vocabulary but it doesn't make reasoning about it in our code very clear.  It has a lexicalAlias `RDF::Vocab::RDAU.P60550.properties[:'http://metadataregistry.org/uri/profile/regap/lexicalAlias']` of "http://rdaregistry.info/Elements/u/extent.en" which is really what we want.

We can add this to our terms so that it can be referred to clearly in our code.