Use markdown to style some of our descriptions
Many "description" or "abstract" fields (at the Item level as well as Communities and Collections) contain HTML tags. Because these are text fields, HTML is not rendered in the UI and text looks garbled and it's way less readable than ideal.

Markdown should work really well for this since that's already used in many of the tools staff working in repositories are familiar with.

Use redcarpet gem because it basically does what we want :)