Fix rdf-n3 lexer error
This change modifies the n3 graph notation to treat the offending
objects as strings rather than elements.

The problem was likely cause with a change in the lexer on rdf-n3 that
is nwo more accurate in how it traverses and checks the elements on the
graph.

This error is not caught in all lexers, that is why the error was not
being shown in earlier versions.