make delegating pagination of collections default
PaginatingDecorator works for all classes, so any Decorator for a collection of a class that needs to be decorated will at some point need to add this same method with the same contents. Preemptively we can make all Decorators behave this way by default to make future work easier with a smart default.