Refactor to a Depositable view
Assuming that depositable items will have
* title and alternative title
* creator(s)
* description
* creation date
* subjects
* item type
* doi
* license

Now the problem is mapping the actual attributes to these things and providing labels.  The decorator is now responsible for this.

Then I was able to remove the item and thesis views.  For convenience I moved the 'more_information' content into the eponymous partials.