Added views for `Digitization::Book`
This is the first draft which basically copied the existing Item views,
helper, exporter and decorator then used the decorator to map similar
but differently named terms. Also added a basic test.

A humanize_uri_or_literal helper method was necessary to lookup subject
values that could be strings or URIs.