Add brakeman to github workflow
Brakeman is a Ruby on Rails static analysis security tool.  We had been running it manually when the gem was bumped but having it in the github workflow will mean that we will have actionable automatic feedback.

There are Github Actions in the marketplace which have more features but this seems like a good first step.