Fix a poor use of tap in seeds
There isn't a good reason to use tap here.  We need to have created the Book before using it (ActiveRecord::RecordNotSaved: You cannot call create unless the parent is saved) so there isn't an intermediate state that we want to operate on.  Also don't need the extra save to the database to create the Fulltext object.

Used the bang version so we'd get failure and feeback from the command line.