Add `Community` and `Collection` objects to reindex task
@henryzhang87 is working on upgrading our Solr to the latest 6.6.  He noticed that our reindex task leaves the Communities and Collections empty.

Turns out we don't have a task to reindex the Communities/Collections.  I add this here.

There was also a bug in the default `desired_batch_size`.  If `args.batch_size` was `nil` then the `desired_batch_size` was assigned 0 (`to_i`) and the update_solr operations were performed 0 times.