Fix uri and predicates
This should be http and without the html prefix.  Confirmed with the original spec and the sample triples.

Previously we changed alt_title to alternative_titles, so we need to use the correct column name.

Also was using the incorrect predicate for rights.  This was found when ingesting the folk fest triples and proven true when revisiting the data dictionary.