Add task that will kick off job for digitization metadata
Initially this was all just one big rake task.  Moving the logic to a Job allows the inputs to be validated and the logic to be tested.

CSV is part of the ruby standard library.  We've adopted a practice of using bundle to manage the require statements.  Since this is part of the standard library and outside of bundler's perview we prefer to include the require statement once in an initializer.