Add task that will kick off setup job for artifacts
Initially this was all just one big rake task.  Moving the logic to a Job allows the inputs to be validated and the logic to be tested.

This is just the setup for the jobs that will actually do the work.  Thinking that each object should have it's own job so that the transactions can stay small and we can get feedback on them independently.