Tasks for Jupiter Preservation (#2567)
* Fix the preserve rake task

The method that this was calling was lost long ago in a refactor.  Updating the method called and fixing a bug when the `batch_size` is `nil`.

* Preserve a subset of our Items and Theses

Modified the existing task to take a `Date` as an argument. The existing `update_on_or_after` scope will get any item/thesis that has been created or updated in that interval.

* Add a task that will clear the preservation queue

This may only be used on one occasion, but currently we have a preservation queue filled  with inoperable keys.  We've repaired for new items.