Add job that will attach artifacts
(high resolution pdf, full text, and link to preservation storage) to digitization objects.

Thinking that each object should have it's own job so that the transactions can stay small and we can get feedback on them independently.

This is a real life aip but I did clip the pdf to one page