Fix situation with error handling
If a book being created had validation errors then the batch ingest status would not be updated because the validation would continue to fail.  So only add the book to the batch ingest after save has been completed.

Add a helper method to quickly display the peel number (peel_id, run, part_number) and custom error message for the validation.