Preserve the `doi` namespace
All our existing doi's have a `doi:` prefix.  This is the output we use for identifiers in OAI.

Rather than make a change that has unknown repercussions we will continue the same doi structure.

I investigated if the Datacite API recognized DOIs in the format `doi:<prefix>/<id>`.  It does not.

I investigated if the Datacite API recognized our existing DOIs in the `<prefix>/<id>` format and it does. We can be assured that this change will ensure continuity of the existing data.