I18n for unique peel_number validation
Couldn't just edit the message in locals/en.yml because the helpful part is showing the uniqueness of the whole scope (peel_id, run, part_number) a.k.a the peel number.  So have to use the proc to gain access to the object so that we can interpoate the peel_number.

This is an improvement if we ever need to translate this application.