positional arguments and keyword arguments changed
`warning: Using the last argument as keyword parameters is deprecated; maybe ** should be added to the call`

In Ruby 2.7, the way it handles the positional arguments and keyword arguments is changed, aiming to provide a smooth transition to Ruby 3.

If you were passing a Hash as a keyword argument, now you'd need to update your code to use the double splat  `**` operator.

https://bloggie.io/@kinopyo/how-to-fix-ruby-2-7-warning-using-the-last-argument-as-keyword-parameters-is-deprecated