Change DraftItem date_created value to string type
DraftItem model needs to accept dates in format 'YYYY-MM-DD',
'YYYY-MM', and 'YYYY'. The current model's behavior does not allow this
due to the date_created value type is date, which only accepts dates
with formats 'YYYY-MM-DD', 'YYYY-MM'. Furthermore the UI needs to change
in order to accept this change.

The date_crated type will change from date to string and we need to add
back and front end validations to handle partial dates.