Add sorted files for item views
Currently the files are shown in the order they finished uploading. This
change will make add a defined way of sorting the files in alphabetical
order to keep the functionality consistent.

Current changes include adding a method for reading files in models
Item, Thesis, DraftItem, and DrafThesis.

Followed the path of adding an extra method to the models as modifying
the files getter would be adding an extra join to the db when sorting
is not needed.