Change depositable's behaviour to attach files
This change allows having a multi-file attachment even inside a base
transaction block.

Before, the behaviour for attaching files in the depositable model
allowed attaching a single file at a time. Now we make sure all of the
copy callbacks are run inside a transaction block, actually copying the
files, and not just the last one.