Disable Spring in GithubActions
I noticed that `rails test` was reporting low coverage (1.46%).  I tried a number of things in the SimpleCov readme without success.  In the process I did discovery that `rake test` was reporting reasonable coverage (82.06%). This lead me to reread the DISABLE_SPRING env var suggestion in the SimpleCov readme which did the trick.  Spring keeps your Rails application preloaded so you don't need to boot it everytime you test, etc.  Not necessary in CI and interferes with SimpleCov which requires that it's loaded before anything else.