WIP Add report generation jobs
This is a first pass at creating the era audit reports through an active
job.

This first approach seems too long as it generates all the reports in
the same job. We also iterate through all Item and Thesis multiple
times.

Should we split the report generation through multiple jobs?